+++ ExampleApp Android Compose/src/main/java/com/example/exampleappandroidcompose/MainActivity.kt
@@ -1,0 +1,28 @@
+package com.example.exampleappandroidcompose
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.ui.Modifier
+import com.example.exampleappandroidcompose.ui.screens.SignUpScreen
+import com.example.exampleappandroidcompose.ui.theme.FlowFormsTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            FlowFormsTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    SignUpScreen()
+                }
+            }
+        }
+    }
+}
