+++ ExampleApp Android Compose/src/main/java/com/example/exampleappandroidcompose/ui/theme/Dimension.kt
@@ -1,0 +1,10 @@
+package com.example.exampleappandroidcompose.ui.theme
+
+import androidx.compose.ui.unit.dp
+
+object Padding {
+    val small = 5.dp
+    val medium = 10.dp
+    val big = 15.dp
+    val huge = 20.dp
+}
