+++ ExampleApp Android Compose/src/main/res/drawable/ic_visibility.xml
