+++ ExampleApp Android Compose/src/main/res/mipmap-xhdpi/ic_launcher.webp
