+++ ExampleApp Android Compose/src/main/res/mipmap-xxhdpi/ic_launcher_round.webp
