+++ ExampleApp Android Compose/src/main/res/mipmap-xxxhdpi/ic_launcher.webp
