+++ ExampleApp Android Compose/src/main/res/values/colors.xml
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+</resources>
