+++ buildSrc/src/main/java/Versions.kt
@@ -16,6 +16,7 @@ object Versions {
     const val LIFECYCLE_COMMON = "2.5.1"
     const val LIFECYCLE_EXTENSIONS = "2.2.0"
     const val GUAVA = "31.1-jre"
+    const val ACTIVITY_COMPOSE = "1.7.2"
 
     //test
     const val COROUTINES_TEST = "1.6.4"
