+++ settings.gradle.kts
@@ -1,3 +1,3 @@
 include(":FlowForms-Core", ":ExampleApp Android", ":shared")
 rootProject.name = "FlowForms"
-include(":exampleappandroidcompose")
+include(":ExampleApp Android Compose")
